# Remove writing through stream when caching data